flask-docker/requirements.txt
beautifulsoup4==4.11.2
Flask==2.2.2
newspaper3k==0.2.8
pandas==1.4.2
requests==2.28.1
selenium==4.8.0
